labm8/py/test: Replace --test_with_gpu with environment var.
This removes the --test_with_gpu flag, and instead checks if
the environment variable TEST_WITH_GPU == 1. This is to remove
the need for parsing FLAGS before calls to the GPU scheduler,
allowing the GPU scheduler to run at the module level of code,
e.g.

        from labm8.py import gpu_scheduler

        gpu_scheduler.LockExclusiveProcessGpuAccess()

As is the recommended usage pattern in the module docs.

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>

[Exported from 2d2a50a3b8f5c539502407cda1dbd47bca70d73a]